wip dump

+++ erigon-lib/state/domain_test.go
@@ -1597,8 +1597,6 @@ func TestDomain_PruneAfterAggregation(t *testing.T) {
 }
 
 func TestDomain_Unwind(t *testing.T) {
-	t.Skip("fix me!")
-
 	db, d := testDbAndDomain(t, log.New())
 	defer d.Close()
 	defer db.Close()
@@ -1620,8 +1618,8 @@ func TestDomain_Unwind(t *testing.T) {
 		var preval1, preval2, preval3, preval4 []byte
 		for i := uint64(0); i < maxTx; i++ {
 			dc.SetTxNum(i)
-			if i%3 == 0 {
-				if i%12 == 0 {
+			if i%3 == 0 && i > 0 { // once in 3 tx put key3 -> value3.i and skip other keys update
+				if i%12 == 0 { // once in 12 tx delete key3 before update
 					err = dc.DeleteWithPrev([]byte("key3"), nil, preval3)
 					require.NoError(t, err)
 					preval3 = nil
@@ -1634,13 +1632,13 @@ func TestDomain_Unwind(t *testing.T) {
 				preval3 = v3
 				continue
 			}
+
 			v1 := []byte(fmt.Sprintf("value1.%d", i))
 			v2 := []byte(fmt.Sprintf("value2.%d", i))
 			nv3 := []byte(fmt.Sprintf("valuen3.%d", i))
 
 			err = dc.PutWithPrev([]byte("key1"), nil, v1, preval1)
 			require.NoError(t, err)
-
 			err = dc.PutWithPrev([]byte("key2"), nil, v2, preval2)
 			require.NoError(t, err)
 			err = dc.PutWithPrev([]byte("k4"), nil, nv3, preval4)
@@ -1725,7 +1723,6 @@ func TestDomain_Unwind(t *testing.T) {
 		})
 		t.Run("WalkAsOf"+suf, func(t *testing.T) {
 			t.Helper()
-			t.Skip()
 
 			etx, err := tmpDb.BeginRo(ctx)
 			defer etx.Rollback()
@@ -1751,7 +1748,12 @@ func TestDomain_Unwind(t *testing.T) {
 		t.Run("HistoryRange"+suf, func(t *testing.T) {
 			t.Helper()
 
-			etx, err := tmpDb.BeginRo(ctx)
+			tmpDb2, expected2 := testDbAndDomain(t, log.New())
+			defer expected2.Close()
+			defer tmpDb2.Close()
+			writeKeys(t, expected2, tmpDb2, unwindTo)
+
+			etx, err := tmpDb2.BeginRo(ctx)
 			defer etx.Rollback()
 			require.NoError(t, err)
 
@@ -1759,16 +1761,54 @@ func TestDomain_Unwind(t *testing.T) {
 			defer utx.Rollback()
 			require.NoError(t, err)
 
-			ectx := expected.MakeContext()
+			ectx := expected2.MakeContext()
 			defer ectx.Close()
 			uc := d.MakeContext()
 			defer uc.Close()
 
-			et, err := ectx.hc.HistoryRange(int(unwindTo)+1, -1, order.Asc, -1, etx)
+			et, err := ectx.hc.HistoryRange(int(unwindTo)-1, -1, order.Asc, -1, etx)
 			require.NoError(t, err)
 
-			ut, err := uc.hc.HistoryRange(int(unwindTo), -1, order.Asc, -1, utx)
+			est, err := ectx.IteratePrefix2(etx, nil, nil, -1)
 			require.NoError(t, err)
+			for est.HasNext() {
+				k, v, err := est.Next()
+				require.NoError(t, err)
+
+				fmt.Printf("estate key %s -> %s\n", string(k), string(v))
+			}
+
+			iet, err := ectx.hc.IdxRange([]byte("key3"), 0, math.MaxInt, order.Asc, -1, etx)
+			require.NoError(t, err)
+			ilist := make([]uint64, 0)
+			for iet.HasNext() {
+				txn, err := iet.Next()
+				require.NoError(t, err)
+				ilist = append(ilist, txn)
+			}
+
+			uet, err := uc.hc.IdxRange([]byte("key3"), 0, math.MaxInt, order.Asc, -1, utx)
+			require.NoError(t, err)
+			ulist := make([]uint64, 0)
+			for uet.HasNext() {
+				txn, err := uet.Next()
+				require.NoError(t, err)
+				ulist = append(ulist, txn)
+			}
+			fmt.Printf("key3 e list %v\n", ilist)
+			fmt.Printf("key3 u list %v\n", ulist)
+
+			ut, err := uc.hc.HistoryRange(int(unwindTo)-1, -1, order.Asc, -1, utx)
+			require.NoError(t, err)
+
+			ust, err := uc.IteratePrefix2(utx, nil, nil, -1)
+			require.NoError(t, err)
+			for ust.HasNext() {
+				k, v, err := ust.Next()
+				require.NoError(t, err)
+
+				fmt.Printf("ustate key %s -> %s\n", string(k), string(v))
+			}
 
 			compareIterators(t, et, ut)
 		})
@@ -1811,12 +1851,12 @@ func TestDomain_Unwind(t *testing.T) {
 	writeKeys(t, d, db, maxTx)
 	//unwindAndCompare(t, d, db, 14)
 	unwindAndCompare(t, d, db, 11)
-	unwindAndCompare(t, d, db, 10)
-	unwindAndCompare(t, d, db, 8)
-	unwindAndCompare(t, d, db, 6)
-	unwindAndCompare(t, d, db, 5)
-	unwindAndCompare(t, d, db, 2)
-	unwindAndCompare(t, d, db, 0)
+	//unwindAndCompare(t, d, db, 10)
+	//unwindAndCompare(t, d, db, 8)
+	//unwindAndCompare(t, d, db, 6)
+	//unwindAndCompare(t, d, db, 5)
+	//unwindAndCompare(t, d, db, 2)
+	//unwindAndCompare(t, d, db, 0)
 
 	return
 }
@@ -1824,6 +1864,24 @@ func TestDomain_Unwind(t *testing.T) {
 func compareIterators(t *testing.T, et, ut iter.KV) {
 	t.Helper()
 
+	for {
+		ek, ev, err1 := et.Next()
+		fmt.Printf("%s %s %v\n", ek, ev, err1)
+		if !et.HasNext() {
+			break
+		}
+	}
+
+	fmt.Printf("unwindedIterhas extra keys\n")
+	i := 0
+	for {
+		uk, uv, err2 := ut.Next()
+		fmt.Printf("i=%d key %s v %s err %v\n", i, string(uk), string(uv), err2)
+		i++
+		if !ut.HasNext() {
+			break
+		}
+	}
 	for {
 		ek, ev, err1 := et.Next()
 		uk, uv, err2 := ut.Next()
